+++ extensions/twilio/v2/actions/getMessages/config/dataPoints.ts
@@ -1,0 +1,16 @@
+import { type DataPointDefinition } from '@awell-health/extensions-core'
+
+export const dataPoints = {
+  numberOfMessages: {
+    key: 'numberOfMessages',
+    valueType: 'number',
+  },
+  allMessages: {
+    key: 'allMessages',
+    valueType: 'strings_array',
+  },
+  latestMessage: {
+    key: 'latestMessage',
+    valueType: 'string',
+  },
+} satisfies Record<string, DataPointDefinition>