+++ extensions/twilio/v2/actions/index.ts
@@ -1,2 +1,3 @@
 export { sendSms } from './sendSms'
 export { sendSmsWithMessagingService } from './sendSmsWithMessagingSid'
+export { getMessages } from './getMessages'